move files around to get app to start with gunicorn
