+++ main.py
@@ -1,0 +1,3 @@
+from app.main import app
+if __name__ == "__main__":
+    app.run(host='0.0.0.0')
