+++ requirements.txt
@@ -1,7 +1,7 @@
 api==0.0.7
 blackfynn==2.8.0
 boto3==1.10.28
-botocore==1.13.28
+botocore==1.13.34
 certifi==2019.11.28
 chardet==3.0.4
 Click==7.0
@@ -9,6 +9,7 @@ docutils==0.15.2
 Flask==1.1.1
 flask-marshmallow==0.10.1
 get==2019.4.13
+gunicorn==20.0.4
 idna==2.8
 itsdangerous==1.1.0
 Jinja2==2.10.3
